requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
asgiref==3.8.1
openai==1.66.2
# selenium and g4f removed — no longer used (FreeProvider uses local Ollama via httpx)
httpx>=0.27.0
google-generativeai==0.8.3
anthropic==0.42.0
aiohttp==3.11.11
pytest==8.3.4
pytest-asyncio==0.24.0